requirements.txt
kaggle
kagglehub
Pillow
tqdm
torch
torchvision
transformers
wandb
scikit-learn
datasets
huggingface-hub
transformers
accelerate
